requirements.txt
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instaloader==4.14
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
pip~=24.3.1
attrs~=23.1.0
wheel~=0.42.0
click~=7.1.2
Brotli~=1.1.0
pytz~=2023.3.post1
setuptools~=69.0.2
packaging~=23.2
Jinja2~=3.1.2
python-dotenv~=1.0.0